dispatcher/requirements.txt
redis==5.0.1
pymongo==4.9.2
motor==3.6.0
python-dotenv==1.0.1
typing-extensions==4.12.2
